+++ delete_sean.php
@@ -1,8 +1,12 @@
 <?php
+include("checks.php");
 require_once 'connect1.php';
 $mysqli = new mysqli($host, $user, $password, $database) or die ("Невозможно подключиться к серверу");
 $id = $_GET['id'];
-$result = $mysqli->query("DELETE FROM sean WHERE id='$id'");
-header("Location: sean.php");
+if ($_SESSION['type'] == 2)
+    $result = $mysqli->query("DELETE FROM sean WHERE id='$id'");
+else
+    echo "Недостаточно прав";
+header("Location: seanAdm.php");
 exit;
 ?>
